Store issue Fixed

+++ app/Http/Controllers/ItemController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use App\Models\Device;
 use App\Models\Item;
+use App\Models\ItemInStore;
 use App\Models\Organization;
 use App\Models\Plan;
 use App\Models\PlanDetail;
@@ -226,12 +227,29 @@ class ItemController extends Controller
     public function delete_item($id)
     {
 
-        $res = Item::where('id', $id)->delete();
+        $item_in_device = Item::where('id', $id)->get();
 
-        if ($res) {
-            return back()->with(['success' => 'One Item is Deleted from system']);
-        } else {
-            return back()->with(['error' => 'Item NOT Deleted from system. Try Again!']);
+        if(count($item_in_device) > 0){
+
+            $state_of_item = false;
+
+            foreach ($item_in_device as $i_in_d) {
+                $check_item_in_store = ItemInStore::where('item_id', '=', $i_in_d->id)->get();
+                $state_of_item = count($check_item_in_store) > 0 ? true : false;
+            }
+
+            if ($state_of_item) {
+                return back()->with(['error' => "SORRY, Can't delete Item. \nMake sure that the item is not existing in a store."]);
+            }
+            
+            $res = Item::where('id', $id)->delete();
+    
+            if ($res) {
+                return back()->with(['success' => 'One Item is Deleted from system']);
+            } else {
+                return back()->with(['error' => 'Item NOT Deleted from system. Try Again!']);
+            }
         }
+
     }
 }
